Add settings tabs and routes
